Improved handling of periodic boundary conditions.
MINOR CHANGE IN DEFAULT BEVAHIOR:

Default handing now looks for the most common periodic image of
contributing atoms, and moves CG site to that place. Added option to
post wrap system across perodic image to facilitate mapping a wrapped
system and maining the wrap. This contrasts to the old version of
simpling choosing the image containing the first atom.

Test added to confirm boundary handling. Numerically, it is less
accurate than others. It is unclear if this is due to VMD (used for
creating references) or the process itself. Charge test modified to
account for differing default behavior (consusus procedure above).